+++ PieCharts/Core/Chart/PieChart.swift
@@ -31,7 +31,22 @@ import UIKit
     @IBInspectable public var animDuration: Double = 0.5
     
     /// Start angle of chart, in degrees, clockwise. 0 is 3 o'clock, 90 is 6 o'clock, etc.
-    @IBInspectable public var referenceAngle: CGFloat = 0
+    @IBInspectable public var referenceAngle: CGFloat = 0 {
+        didSet {
+            for layer in layers {
+                layer.clear()
+            }
+            
+            let delta = (referenceAngle - oldValue).degreesToRadians
+            for slice in slices {
+                slice.view.angles = (slice.view.startAngle + delta, slice.view.endAngle + delta)
+            }
+            
+            for slice in slices {
+                slice.view.present(animated: false)
+            }
+        }
+    }
     
     var animated: Bool {
         return animDuration > 0
@@ -41,7 +56,7 @@ import UIKit
     
     public fileprivate(set) var container: CALayer = CALayer()
     
-    fileprivate var slices: [PieSlice] = []
+    public fileprivate(set) var slices: [PieSlice] = []
     
     public var models: [PieSliceModel] = [] {
         didSet {
@@ -116,7 +131,9 @@ import UIKit
         slice.view.referenceAngle = referenceAngle.degreesToRadians
         
         slice.view.sliceDelegate = self
-     
+
+        self.delegate?.onGenerateSlice(slice: slice)
+
         return (newEndAngle, slice)
     }
     
@@ -205,6 +222,15 @@ import UIKit
         slices = []
     }
     
+    public func clear() {
+        for layer in layers {
+            layer.clear()
+        }
+        layers = []
+        models = []
+        removeSlices()
+    }
+    
     open override func prepareForInterfaceBuilder() {
         animDuration = 0
         strokeWidth = 1
@@ -215,6 +241,11 @@ import UIKit
         }
         
         self.models = models
+    }
+    
+    open override func layoutSubviews() {
+        super.layoutSubviews()
+        container.frame = bounds
     }
 }